ConversionProgram
This program asks the user to pick which conversion they want to make and then asks them the information that they want to convert. For example if the user were to pick 'A', the program would  ask them the quantity in pounds and then it would print out the conversion.